Add specialized support for generating NNCPs for static routes
Users can pass in a list of destinations and the component will generate
a NNCP which removes destinations which are prefixed with `~` and adds
all other destinations.

Note that this is intentionally the same format that component-cilium
supports for configuring custom destination CIDRs for egress policies,
so that users can manage a single list of destinations which is suitable
both for this component's `staticRoutes` parameter and for
component-cilium's egress IP range `destination_cidrs` parameter.