Implement workaround for nmstate-handler pods stuck in `Pending`
We deploy an Espejote ManagedResource which watches the nmstate-handler
pods and deletes any pods which are stuck in `Pending` with reason
`Unschedulable`.

This will delete pods that get created with impossible node selectors
immediately after the operator reconciles the namespace (which removes
the `openshift.io/node-selector=""` annotation due to a bug in the
operator).

Usually, ArgoCD will reconfigure the node selector annotation fairly
quickly, but when the OLM updates the operator the namespace is
reconciled (removing the node selector annotation) and the DaemonSet is
updated with a new container image which creates a window where a few
DaemonSet pods will be created before ArgoCD can reconfigure the node
selector annotation.